Revert "refactor(design): remove redundant observable.ts and its related meth…"
This reverts commit bbe14427d72388a8aef794964ce7a2ab5faa475e.
