+++ packages/design/src/index.ts
@@ -41,6 +41,7 @@ export { type ISelectListProps, SelectList } from './components/select-list';
 export { type ISliderProps, Slider } from './components/slider';
 export { type ITooltipProps, Tooltip } from './components/tooltip';
 export { type ITreeNodeProps, type ITreeProps, Tree, TreeSelectionMode } from './components/tree';
+export { useObservable } from './hooks/observable';
 export { enUS, zhCN } from './locale';
 export { type ILocale } from './locale/interface';
 export { defaultTheme, greenTheme, themeInstance } from './themes';
