+++ packages/sheets-ui/src/views/count-bar/ZoomSlider.tsx
@@ -21,8 +21,7 @@ import {
     IContextService,
     IUniverInstanceService,
 } from '@univerjs/core';
-import { useObservable } from '@univerjs/ui';
-import { Slider } from '@univerjs/design';
+import { Slider, useObservable } from '@univerjs/design';
 import { SetWorksheetActivateCommand } from '@univerjs/sheets';
 import { useDependency } from '@wendellhu/redi/react-bindings';
 import React, { useEffect, useState } from 'react';
