+++ packages/ui/src/components/hooks/observable.ts
@@ -15,87 +15,41 @@
  */
 
 import type React from 'react';
-import { useEffect, useRef, useState } from 'react';
+import { useRef, useState } from 'react';
 import type { Observable, Subscription } from 'rxjs';
 
-type ObservableOrFn<T> = Observable<T> | (() => Observable<T>);
-
-function unwrap<T>(o: ObservableOrFn<T>): Observable<T> {
-    if (typeof o === 'function') {
-        return o();
-    }
-
-    return o;
-}
-
-function showArrayNotEqual(arr1: unknown[], arr2: unknown[]): boolean {
-    if (arr1.length !== arr2.length) {
-        return true;
-    }
-
-    return arr1.some((value, index) => value !== arr2[index]);
-}
-
-export function useObservable<T>(observable: ObservableOrFn<T>, shouldHaveSyncValue: true): T;
+export function useObservable<T>(observable: Observable<T>, defaultValue?: T): T | undefined;
 export function useObservable<T>(observable: Observable<T>, defaultValue?: T, shouldHaveSyncValue?: true): T;
 export function useObservable<T>(
-    observable: ObservableOrFn<T>,
-    shouldHaveSyncValue?: false,
+    observable: Observable<T>,
     defaultValue?: T,
-    deps?: any[]
-): T | undefined;
-export function useObservable<T>(
-    observable: ObservableOrFn<T>,
-    shouldHaveSyncValue?: boolean,
-    defaultValue?: T,
-    deps?: any[]
+    shouldHaveSyncValue?: true
 ): T | undefined {
     const observableRef = useRef<Observable<T> | null>(null);
     const subscriptionRef = useRef<Subscription | null>(null);
-    const depsRef = useRef<any[] | undefined>(deps ?? undefined);
-    const initializedRef = useRef<boolean>(false);
 
     let setValue: React.Dispatch<React.SetStateAction<T | undefined>>;
+    let initialized = false;
     let innerDefaultValue: T | undefined;
-
-    const shouldResubscribe = (() => {
-        if (typeof depsRef.current !== 'undefined') {
-            const _deps = deps ?? [];
-            if (showArrayNotEqual(depsRef.current, _deps)) {
-                depsRef.current = _deps;
-                return true;
-            }
-
-            return false;
-        }
-
-        return observableRef.current !== observable;
-    })();
-
-    if (!subscriptionRef.current || shouldResubscribe) {
-        observableRef.current = unwrap(observable);
+    if (!observableRef.current || observableRef.current !== observable) {
         subscriptionRef.current?.unsubscribe();
-        subscriptionRef.current = observableRef.current.subscribe((value) => {
+        subscriptionRef.current = observable.subscribe((v) => {
             if (setValue) {
-                setValue(value);
+                setValue(v);
             } else {
-                innerDefaultValue = value;
-                initializedRef.current = true;
+                innerDefaultValue = v;
+                initialized = true;
             }
         });
     }
 
-    if (shouldHaveSyncValue && !initializedRef.current) {
+    if (shouldHaveSyncValue && !initialized) {
         throw new Error('[useObservable]: expect shouldHaveSyncValue but not getting a sync value!');
     }
 
     const s = useState<T | undefined>(innerDefaultValue || defaultValue);
     const value = s[0];
     setValue = s[1];
-
-    useEffect(() => {
-        return () => subscriptionRef.current?.unsubscribe();
-    }, []);
 
     return value;
 }
